requirements.txt
# $ pip install -r requirements.txt

digi-xbee~=1.5.0
pyqtgraph~=0.13.7
PySide6~=6.8.0.2
